// src/app/models/msmagasindomains/panier/commande-produit.model.ts
export class CommandeProduit {
    key?: string;
    code: string;
    designation: string;
    prixUnitaire: number;
    descriptionProduit: string;
    zoneGeographiqueId: number;
    isValid:boolean;
    image: string;
    fidBoutique: string;
    fidProprietaire: string;
    fidAcheteur: string;
    quantite:string;
    commandId:string;
}